Fix client usage estimation for relays
There are a number of problems with trying to compute client usage
costs just by looking at the payments endpoint.

The old code here used to account money being paid to a relay as
money paid for client usage. Which resulted in ridiculous values
of course.

Now we properly filter down to just payments we have sent. Although
this makes usage files non-portable.