More fixes to relay accounting display
This is yet another example of why a single function to process and
display relay and client usage is just a mistake waiting to happen.

For relays traffic is passed through so upload is almost always equal
to download and you're only paid for the forwarding. On the other hand
as a user you pay seperately for upload and download and it makes quite
a bit of sense to sum them for the total usage.

Not only did this slip my attention but I also missed that service fees
where being *added* to relay usage because they where being added to the
total bill for clients. Another dumb avoidable mistake.

The only real solution to this is to rebuild the backend and have it do
all the required processing from scratch.