Update dependency tree
This started with updating node-sass to play nice with Python 3.10 it
spiraled out from there based on how dependencies of dependencies
conflicted. The new major change here is an update to prettier which
creates a bunch of new warnings based on new lints. I have reduced these
from errors, the previous lint configuration in the same of keeping this
change sane in scope.