Add config for svgr
This patch adds the config file for a svgr a transitive dep that
typically works fine outwith a config. But in the openwrt firmware
builder it will continue going up a directory looking for it's config
until it eventually confuses the .config file in openwrt for some sort
of config directory, causing an error.