Move reset exit to advanced settings page
Users have been pushing this button whenever their connection was
disrupted, becuase it's the most obvious error and there are no warnings
attached. Sadly this just ends up with the user even more confused as
they can't re-register to the exit easily.

This patch moves the button to the advanced settings page and adds
specific warnings around using it.