Update port orderings for ipq40xx DSA transition
The ipq40xx arch in openwrt is now being transitioned to DSA. In order
to faciliate this I've added DSA port lists for ipq40xx models and
changed the way the kernel version check works to try on all routers
rather than only an explicit list of DSA supported ones. As more
architecutres move to DSA we can hopefully even eliminate the port
orderings as a concept.