Better form handling for install details
Previously the install details card submission button could fail silently
which was causing confusion for data syncing to ops. This change is the
frontend side to an update for this endpoint which now returns a detailed
error message if something fails due to invalid inputs. Also changed are
the data types to line up more closely with what the backend expects for
Option types versus non-optional strings which I suspect was also part of
the silent failures.